Refactor Moved request functions to spotify_requests.py
