+++ src/cli.py
@@ -1,5 +1,4 @@
-import requests
-import os
+from src.spotify_requests import search_tracks, search_albums, search_artists, search_genres
 
 welcome = True
 
@@ -22,46 +21,3 @@ def cli():
         print(e)
         welcome = not welcome
         cli()
-
-
-def search_tracks():
-    track_name = input("Enter the track name: ")
-    access_token = os.getenv("SPOTIFY_CLIENT_TOKEN")
-
-    query = track_name.replace(' ', '+')
-
-    search_url = "https://api.spotify.com/v1/search"
-
-    headers = {
-        "Authorization": f"Bearer {access_token}"
-    }
-
-    params = {
-        "q": query,
-        "type": "track",
-        "limit": 10
-    }
-
-    response = requests.get(search_url, headers=headers, params=params)
-
-    if response.status_code == 200:
-        # Successful API call
-        track_data = response.json()['tracks']['items'][0]
-        print(track_data)
-        return track_data
-    else:
-        # API call failed
-        print(f"Failed to fetch data: {response.status_code}")
-        return None
-
-
-def search_albums():
-    print("Searching for albums...")
-
-
-def search_artists():
-    print("Searching for artists...")
-
-
-def search_genres():
-    print("Searching for genres...")
